Add some missing IPA finals
Based on the Wikipedia pinyin article. Now each of the `Ting.all_syllables` can
convert to IPA. Looking at the existing syllables it seems there are quite a few
where we are using /i/ where perhaps we should be using /j/. (or introduce an
ipa_standalone which has the medial as /j/?) Might be worth comparing with some
existing pinyin/ipa tables.

Also added two "?" in the finals table for standalone ㄝ/ê for wade giles and
Palladius. This is only relevant for certain interjections and I'm not sure
WG/Palladius really has anything for this case.